ENH: update events
